separate components by folders
