+++ src/components/AddCompetitionBox/AddCompetitionBox.jsx
