+++ src/components/AddCompetitionBox/AddCompetitionModal.jsx
