+++ src/components/CountdownTimer/DeleteCompetitionButton.jsx
